+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - Update FastSpell to `0.4`.
+- Call FastSpell only one time when `--add_lang_ident`
 
 ## v1.2 2022-11-14
 ### Changed